180511 마법 구현 완료, 리팩토링
1. 마법 추가 완료
 - 마법의 공격범위와 사용할 수 있는 범위
 - 마법의 데미지

2. 리네이밍
- 메시지:  attackPoint -> damage
- 캐릭터: attackedPoint -> damage

3. idle 상태 수정
- 캐릭터가 죽어있으면 return 시키고 있었음
- 공격받으면 defense상태로 변경되서 죽어있는데도 불구하고 상태가 변경됨
- dead 상태에서 상태를 변경하지 않게 바꾸고 죽어있으면 업데이트를 하지 않게 변경

4.수정해야 할 것
- Skill 클래스의 쓸데없는 인자 제거해야 함.

+++ 2D Tile-Based Game/2D Tile-Based Game/ComponentSystem.cpp
@@ -95,7 +95,7 @@ Component* ComponentSystem::FindComponentInRange(Component* mapComponent, Compon
 					//죽은대상이면 검사 제외
 					if (component->IsLive())
 					{
-						for (int i = 0; i < compareTypeList.size(); i++)
+						for (size_t i = 0; i < compareTypeList.size(); i++)
 						{
 							if (compareTypeList[i] == component->GetType())
 							{
